refactor : VO 에서 로직이 추가되므로 엔티티로 승격
