+++ src/main/java/carracing/model/RacingDisplayVO.java
@@ -1,4 +1,4 @@
-package carracing.present;
+package carracing.model;
 
 import java.util.ArrayList;
 import java.util.List;
